77% completed
Sigh, this is harder than I expected, but if someday I am to have my name printed on my own cs textbook, why not practice and make some kick-ass course material for other professors to teach? I was taken aback by a remark from one of the advanced level c# professor "what are you going to do with all that money (becoming an adjunct assistant)" ... you don't do it for the money; you do it to become leet: You do it for the backdoors most expertly woven, the zero-day on one saw coming, and the data dumps that revile the true behind power.